Allow provisioning VMs in stopped state
This allows us to avoid race-conditions between the provisioning of the
private network interface and the VMs already booting. Instead of
power-cycling the control plane and worker node after they're
provisioned we now simply power them on using Terraform in a second
apply step.

This also changes the network interface name for the second interface
from `ens6` to `ens4` on the RHCOS VMs booted from the upstream base
image.