Adjust commit-hieradata.sh to work correctly for full installation process
The script covered a few cases which don't appear in the regular
installation process, but broke in a few cases which appear in the
regular process.

Mainly, we need to ensure that gitfile_checkout doesn't choke on missing
tracking branches for our feature branch, so we always set the feature
branch to track `origin/master`.