Never amend existing commits in commit-hieradata.sh
This commit changes the commit-hieradata.sh script to never amend
existing commits, as that led to a pretty chaotic history in the
hieradata repository combined with the change to make the feature branch
track `origin/master`.

With the current install instructions it's also never necessary to amend
an existing commit, as we merge the changes every time between Terraform
runs.