Implement systemd unit to wipe storagepool partition on first boot
We've observed issues with Ceph failing to initialize the devices on
fresh Exoscale VMs.

This commit deploys a systemd unit which only runs at first boot
(conditions copied from machine-config-daemon-firstboot.service) and
zeroes the storagepool partition.

Most likely, it would be sufficient to zero the first couple GB of the
partition to remove any data that might confuse `ceph-volume` while
setting up the backing storage for the OSD. However, we err on the safe
side and zero the full partition.