Only support custom disk sizes for worker nodes
There's generally no reason to provision extra partitions on masters and
infra nodes, so we remove support for configuring a "data disk" for
those nodes. It's still possible to increase the root disk size via the
global parameter.

Additionally, we change the variable controlling the worker disk size to
explicitly control the "data disk" size.