Provide location for MR URL file to commit script
The `commit-hieradata.sh` script previously had hard-coded location
`/tf/.mr_url.txt` to write the URL of the hieradata MR, if any.

This doesn't work in GitLab CI, as the directory structure looks
different to the structure we have during cluster setup.

This commit extends the `commit-hieradata.sh` script to take the full
path to the `.mr_url.txt` file as a second parameter on the command
line and updates the Terraform code to provide the path to the script.

With this change, we can use Terraform's `${path.cwd}` to always create
the `.mr_url.txt` file in Terraform's working directory.