Switch infra nodes to non-instancepool by default
We've realized that with the current state of NLBs (which aren't
suitable for replacing the Puppet-managed LBs at this point), having an
instancepool for the infra nodes brings more issues than it solves.

This commit introduces a new variable `infra_use_instancepool` which
defaults to `false` and must be set to `true` in addition to
`use_instancepools = true` for the module to provision an instancepool
for the infra nodes.